pyproject.toml
[tool.poetry]
name = "fastapi-django"
package-mode = false
version = "0.1.0"

# doc
authors = ["Weiliang Li <to.be.impressive@gmail.com>"]
description = "fastapi and django integration"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"

# 3rd party
django = "^5.0"
fastapi = "^0.114.2"
orjson = "^3.10.7"
pydantic = "^2.8.2"
uvicorn = "^0.30.4"
uvloop = {version = "^0.20.0", markers = 'sys_platform != "win32"'}

# optional
psycopg2-binary = "^2.9.5"
# you don't need whitenoise if you mount the django APP with FastAPI
whitenoise = "^6.7.0"
pyjwt = "^2.9.0"

[tool.poetry.group.dev.dependencies]
django-stubs = {extras = ["compatible-mypy"], version = "^5.0"}
ipython = "^8.26.0"
mypy = "^1.11.1"
ruff = "^0.6.5"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool.isort]
multi_line_output = 3
profile = "black"
